remove promise return from static register call
